+++ src/main/java/dev/lambdaurora/auroraslanterns/AurorasLanternsMixinPlugin.java
@@ -35,17 +35,16 @@ public final class AurorasLanternsMixinPlugin implements IMixinConfigPlugin {
 
 	@Override
 	public String getRefMapperConfig() {
-		return "";
+		return null;
 	}
 
 	@Override
 	public boolean shouldApplyMixin(String targetClassName, String mixinClassName) {
-		return this.conditionalMixins.getOrDefault(targetClassName, true);
+		return this.conditionalMixins.getOrDefault(mixinClassName, true);
 	}
 
 	@Override
 	public void acceptTargets(Set<String> myTargets, Set<String> otherTargets) {
-
 	}
 
 	@Override
@@ -55,11 +54,9 @@ public final class AurorasLanternsMixinPlugin implements IMixinConfigPlugin {
 
 	@Override
 	public void preApply(String targetClassName, ClassNode targetClass, String mixinClassName, IMixinInfo mixinInfo) {
-
 	}
 
 	@Override
 	public void postApply(String targetClassName, ClassNode targetClass, String mixinClassName, IMixinInfo mixinInfo) {
-
 	}
 }